# remote-signing/docs/swagger.yaml
definitions:
  server.ErrorResponse:
    description: Standard error response structure
    properties:
      error:
        description: Error message
        example: Invalid request format
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /sign/{uuid}:
    get:
      description: Get the web-based signing interface for a specific signing request
      parameters:
        - description: Signing request UUID
          in: path
          name: uuid
          required: true
          type: string
      produces:
        - text/html
      responses:
        '200':
          description: HTML signing form
          schema:
            type: string
        '404':
          description: Signing request not found
          schema:
            $ref: '#/definitions/server.ErrorResponse'
      summary: Get signing form
      tags:
        - Web Interface
...
swagger: '2.0'